requirements.txt
Flask
python-dotenv
pyhtcc